refactor(test): moved test directory and updated references
